Cosas
like

+++ server/server.js
@@ -1,36 +1,56 @@
 const express = require('express');
-const bodyParser = require('body-parser');
 const connection = require('./db');
 const cors = require('cors');
+const bcrypt = require('bcrypt');
 
 const app = express();
 app.use(cors());
 app.use(express.json());
 
-//codigo para el form del paciente
+const saltRounds = 10;
+
+// Código para el formulario del paciente
 app.post('/api/patients', (req, res) => {
   const { name, paternalLastName, maternalLastName, age, gender, status, email, password } = req.body;
+  const role = 'patient';
 
-  connection.query(
-    'INSERT INTO patients (name, paternalLastName, maternalLastName, age, gender, status, email, password) VALUES (?, ?, ?, ?, ?, ?, ?, ?)',
-    [name, paternalLastName, maternalLastName, age, gender, status, email, password],
-    (error, results) => {
-      if (error) {
-        res.status(500).json({ error });
-      } else {
-        res.status(201).json({ id: results.insertId });
-      }
+  bcrypt.hash(password, saltRounds, (err, hashedPassword) => {
+    if (err) {
+      res.status(500).json({ error: err });
+    } else {
+      connection.query(
+        'INSERT INTO users (email, password, role) VALUES (?, ?, ?)',
+        [email, hashedPassword, role],
+        (error, results) => {
+          if (error) {
+            res.status(500).json({ error });
+          } else {
+            const userId = results.insertId;
+
+            connection.query(
+              'INSERT INTO patients (user_id, name, paternalLastName, maternalLastName, age, gender, status) VALUES (?, ?, ?, ?, ?, ?, ?)',
+              [userId, name, paternalLastName, maternalLastName, age, gender, status],
+              (error, results) => {
+                if (error) {
+                  res.status(500).json({ error });
+                } else {
+                  res.status(201).json({ id: results.insertId });
+                }
+              }
+            );
+          }
+        }
+      );
     }
-  );
+  });
 });
 
-//Codigo y conexion para inciar sesion
-
+// Código y conexión para iniciar sesión
 app.post('/api/login', (req, res) => {
   const { email, password } = req.body;
 
   connection.query(
-    'SELECT * FROM patients WHERE email = ?',
+    'SELECT * FROM users WHERE email = ?',
     [email],
     (error, results) => {
       if (error) {
@@ -39,19 +59,36 @@ app.post('/api/login', (req, res) => {
         if (results.length === 0) {
           res.status(404).json({ message: 'Email no encontrado' });
         } else {
-          const patient = results[0];
-          if (patient.password !== password) {
-            res.status(401).json({ message: 'Contraseña incorrecta' });
-          } else {
-            res.status(200).json({ message: 'Inicio de sesión exitoso', patient });
-          }
+          const user = results[0];
+
+          bcrypt.compare(password, user.password, (err, result) => {
+            if (err) {
+              res.status(500).json({ message: 'Error interno del servidor' });
+            } else {
+              if (result) {
+                connection.query(
+                  'SELECT * FROM patients WHERE user_id = ?',
+                  [user.id],
+                  (error, results) => {
+                    if (error) {
+                      res.status(500).json({ message: 'Error interno del servidor' });
+                    } else {
+                      res.status(200).json({ message: 'Inicio de sesión exitoso', user, patient: results[0] });
+                    }
+                  }
+                );
+              } else {
+                res.status(401).json({ message: 'Contraseña incorrecta' });
+              }
+            }
+          });
         }
       }
     }
   );
 });
 
-//Conexion de la tabla del home
+// Conexión de la tabla del home
 app.get('/api/patients', (req, res) => {
   console.log('GET request to /api/patients received');
   connection.query('SELECT * FROM patients', (error, results) => {
@@ -63,15 +100,33 @@ app.get('/api/patients', (req, res) => {
   });
 });
 
-// eliminar un paciente
-app.delete('/api/patients/:id', (req, res) => {
+// Eliminar un paciente
+app.delete('/api/patients/:id', (req,   res) => {
   const { id } = req.params;
 
-  connection.query('DELETE FROM patients WHERE id = ?', [id], (error, results) => {
+  connection.query('SELECT user_id FROM patients WHERE id = ?', [id], (error, results) => {
     if (error) {
       res.status(500).json({ message: 'Error interno del servidor' });
     } else {
-      res.status(200).json({ message: 'Paciente eliminado exitosamente' });
+      if (results.length === 0) {
+        res.status(404).json({ message: 'Paciente no encontrado' });
+      } else {
+        const userId = results[0].user_id;
+
+        connection.query('DELETE FROM patients WHERE id = ?', [id], (error, results) => {
+          if (error) {
+            res.status(500).json({ message: 'Error interno del servidor' });
+          } else {
+            connection.query('DELETE FROM users WHERE id = ?', [userId], (error, results) => {
+              if (error) {
+                res.status(500).json({ message: 'Error interno del servidor' });
+              } else {
+                res.status(200).json({ message: 'Paciente eliminado exitosamente' });
+              }
+            });
+          }
+        });
+      }
     }
   });
 });
@@ -80,3 +135,4 @@ const PORT = process.env.PORT || 3001;
 app.listen(PORT, () => {
   console.log(`Server iniciado en el puerto ${PORT}`);
 });
+
